updated contact link structure
