+++ pages/contact-us/contact-canada-pension-plan.js
@@ -1,29 +1,39 @@
 import PropTypes from 'prop-types'
 import { Heading, TableContent } from '@dts-stn/service-canada-design-system'
 import { Fragment } from 'react'
-import en from '../locales/en'
-import fr from '../locales/fr'
-import ContactSection from '../components/contact/ContactSection'
-import ContactProvince from '../components/contact/ContactProvince'
-import { getBetaBannerContent } from '../graphql/mappers/beta-banner-opt-out'
-import { getBetaPopupExitContent } from '../graphql/mappers/beta-popup-exit'
-import { getBetaPopupNotAvailableContent } from '../graphql/mappers/beta-popup-page-not-available'
-import { getContactOldAgeSecurityContent } from '../graphql/mappers/contact-old-age-security'
-import logger from '../lib/logger'
+import en from '../../locales/en'
+import fr from '../../locales/fr'
+import ContactSection from '../../components/contact/ContactSection'
+import ContactProvince from '../../components/contact/ContactProvince'
+import { getBetaBannerContent } from '../../graphql/mappers/beta-banner-opt-out'
+import { getBetaPopupExitContent } from '../../graphql/mappers/beta-popup-exit'
+import { getBetaPopupNotAvailableContent } from '../../graphql/mappers/beta-popup-page-not-available'
+import { getContactCanadaPensionPlan } from '../../graphql/mappers/contact-canada-pension-plan'
+import logger from '../../lib/logger'
 import React from 'react'
 
-export default function ContactOldAgeSecurity(props) {
+export default function ContactCanadaPensionPlan(props) {
   /* istanbul ignore next */
   const t = props.locale === 'en' ? en : fr
+
+  const [openModalWithLink, setOpenModalWithLink] = React.useState({
+    isOpen: false,
+    activeLink: '/',
+  })
 
   return (
     <div
       id="homeContent"
-      data-testid="homeContent-test"
-      data-cy="oasContactUsContent"
+      data-testid="contactCPP-test"
+      data-cy="eIContactUsContent"
     >
       <Heading id="my-dashboard-heading" title={props.pageContent.title} />
-      <div className="py-5" />
+      <div
+        className="py-5"
+        data-testid={`${
+          props.pageContent.items.length > 0 && 'tableOfContents-test'
+        }`}
+      />
       <TableContent
         sectionList={props.pageContent.items.map((item, i) => {
           return { name: item.title, link: `#${item.id}` }
@@ -39,6 +49,10 @@ export default function ContactOldAgeSecurity(props) {
           )}
         </Fragment>
       ))}
+
+      {/*  */}
+
+      {/*  */}
     </div>
   )
 }
@@ -69,12 +83,12 @@ export async function getStaticProps({ res, locale }) {
   /* istanbul ignore next */
   const langToggleLink =
     locale === 'en'
-      ? '/fr/contact-old-age-security'
-      : '/contact-old-age-security'
+      ? '/fr/contact-canada-pension-plan'
+      : '/contact-canada-pension-plan'
 
   const t = locale === 'en' ? en : fr
 
-  const pageContent = await getContactOldAgeSecurityContent().catch((error) => {
+  const pageContent = await getContactCanadaPensionPlan().catch((error) => {
     logger.error(error)
     // res.statusCode = 500
     throw error
@@ -82,11 +96,11 @@ export async function getStaticProps({ res, locale }) {
 
   const breadCrumbItems =
     locale === 'en'
-      ? pageContent.en.breadcrumb.map(({ link, text }) => {
-          return { text, link }
+      ? pageContent.en.breadcrumb?.map(({ link, text }) => {
+          return { text, link: '/' + link }
         })
-      : pageContent.fr.breadcrumb.map(({ link, text }) => {
-          return { text, link }
+      : pageContent.fr.breadcrumb?.map(({ link, text }) => {
+          return { text, link: '/' + link }
         })
 
   // const breadCrumbItems = [
@@ -103,13 +117,13 @@ export async function getStaticProps({ res, locale }) {
   /* Place-holder Meta Data Props */
   const meta = {
     data_en: {
-      title: 'My Service Canada Account - Contact Old Age Security',
+      title: 'My Service Canada Account - Contact Canada Pension Plan',
       desc: 'English',
       author: 'Service Canada',
       keywords: '',
     },
     data_fr: {
-      title: 'Mon dossier Service Canada - Contactez Old Age Security',
+      title: 'Mon dossier Service Canada - Régime de Pensions du Canada',
       desc: 'Français',
       author: 'Service Canada',
       keywords: '',
@@ -129,7 +143,7 @@ export async function getStaticProps({ res, locale }) {
   }
 }
 
-ContactOldAgeSecurity.propTypes = {
+ContactCanadaPensionPlan.propTypes = {
   /**
    * current locale in the address
    */
